# swagger.yaml
openapi: 3.0.0
info:
  title: NusaCall API
  description: >-
    API Documentation for NusaCall — the WhatsApp Business Calling backend for Nusanet.
    NusaCall owns calling (signaling, media, routing) and its own user identity (User/
    Organization) — auth no longer depends on nusawa. nusawa remains the source of truth
    for contacts only; the frontend never talks to nusawa directly, every nusawa call is
    relayed through this backend. See docs/INTEGRATION-NUSAWA.md.
    Language is detected from the `Accept-Language` header (`en` default, `id` supported)
    and echoed back via the `Content-Language` response header on every response.
  version: 1.0.0
servers:
  - url: http://localhost:4100/api
    description: Local development server
  - url: /api
    description: Relative API server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "NusaCall-issued JWT access token. Expires in 8 hours."
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: >-
        Preferred response language. Supported: `en` (default), `id`.
        Missing or unsupported values fall back to `en`.
      schema:
        type: string
        enum: [en, id]
        default: en
      example: id
  headers:
    ContentLanguage:
      description: Resolved response language (`en` or `id`), echoed back from `Accept-Language`.
      schema:
        type: string
        enum: [en, id]
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      description: >-
        The user's Nusawork email/password. Validated server-side against Nusawork's
        OAuth2 password grant — never called from the browser. Identity itself (who
        this user is) is sourced from NusaCall's own User table, not Nusawork.
      properties:
        email: { type: string, format: email, example: "agent@nusa.id" }
        password: { type: string, example: "secret123" }
    GoogleLoginRequest:
      type: object
      required: [idToken]
      properties:
        idToken: { type: string, description: "Google ID token from the frontend's Google Sign-In flow." }
...
paths:
...
  /auth/login:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    post:
      tags: [Auth]
      summary: Login with the user's Nusawork email and password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Logged in successfully
          headers:
            Content-Language: { $ref: "#/components/headers/ContentLanguage" }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthResponse" }
        "401":
          description: User not registered locally, account is inactive, or the Nusawork password is invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }

  # ── Google Login ─────────────────────────────────────────────────────────
  /auth/google:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    post:
      tags: [Auth]
      summary: Login with a Google ID token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GoogleLoginRequest" }
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthResponse" }
        "400":
          description: No local User registered for this Google account's email, or the account is inactive
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }

  # ── Refresh Token ────────────────────────────────────────────────────────
  /auth/refresh:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    post:
      tags: [Auth]
      summary: Exchange a refresh token for a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshTokenRequest" }
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthResponse" }
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }

  # ── Logout ───────────────────────────────────────────────────────────────
  /auth/logout:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    post:
      tags: [Auth]
      summary: Logout the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BaseResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ── Current User ─────────────────────────────────────────────────────────
  /auth/me:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    get:
      tags: [Auth]
      summary: Get the currently authenticated user's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ── User Roster ──────────────────────────────────────────────────────────
  /user:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    get:
      tags: [User]
      summary: Get all users (paginated)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
        - name: q
          in: query
          description: Search by name or email
          schema: { type: string }
        - name: isActive
          in: query
          schema: { type: string, enum: ["true", "false"] }
        - name: sortBy
          in: query
          schema: { type: string, enum: [name, email, isActive, createdAt] }
        - name: order
          in: query
          schema: { type: string, enum: [ASC, DESC], default: DESC }
      responses:
        "200":
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/User" }
                      meta: { $ref: "#/components/schemas/PaginationMeta" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    post:
      tags: [User]
      summary: Create a user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "400":
          description: Email already in use
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }
  /user/options:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    get:
      tags: [User]
      summary: Lightweight active-user picker (autocomplete), capped at 50 results
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 50 }
      responses:
        "200":
          description: User options retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/User" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /user/me:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    get:
      tags: [User]
      summary: Get the currently authenticated user's profile (alias of /auth/me), with live presence
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /user/available:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    get:
      tags: [User]
      summary: Get users currently online (live WebSocket connection) and free to take a call
      description: >-
        There is no eligibility flag anymore — every user can receive calls.
        A user appears here purely based on live presence (connected, not
        already on a call), not any stored roster setting.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Available users retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/User" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /user/{id}:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [User]
      summary: Get a single user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: User not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    put:
      tags: [User]
      summary: Update a user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserRequest" }
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "400":
          description: Email already in use
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: User not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }
    delete:
      tags: [User]
      summary: Soft-delete a user (anonymizes email, deactivates)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User deleted successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BaseResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: User not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ── Contact List (read-only proxy over nusawa) ──────────────────────────
  /contact:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    get:
      tags: [Contact]
      summary: Get the agent's contact list (paginated), proxied from nusawa
      description: >-
        Read-only. NusaCall owns no contact data of its own — this relays to nusawa's
        GET /api/contacts using the agent's own nusawa session, captured at login time.
        Results are cached in-memory per agent for a short TTL (NUSAWA_CONTACT_CACHE_TTL,
        default 30s). If the cached nusawa session has expired, the agent must log in
        again — there is no password on hand to silently re-authenticate with.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
        - name: search
          in: query
          description: Search by contact name or phone number
          schema: { type: string }
      responses:
        "200":
          description: Contacts retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Contact" }
                      meta: { $ref: "#/components/schemas/PaginationMeta" }
        "401":
          description: Unauthorized, or the cached nusawa session has expired (re-login required)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "503":
          description: nusawa (contact directory) is currently unreachable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ── Call History ─────────────────────────────────────────────────────────
  /call:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    get:
      tags: [Call]
      summary: Get call history (paginated, filterable)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
        - name: q
          in: query
          description: Search by wa_id, contact name, or wacid
          schema: { type: string }
        - name: status
          in: query
          description: Comma-separated statuses, e.g. "missed,rejected"
          schema: { type: string }
        - name: direction
          in: query
          schema: { type: string, enum: [inbound, outbound] }
        - name: agentEmail
          in: query
          schema: { type: string }
        - name: phoneNumberId
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: sortBy
          in: query
          schema: { type: string, enum: [createdAt, answeredAt, endedAt, durationSeconds] }
        - name: order
          in: query
          schema: { type: string, enum: [ASC, DESC], default: DESC }
      responses:
        "200":
          description: Calls retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Call" }
                      meta: { $ref: "#/components/schemas/PaginationMeta" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /call/stats:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    get:
      tags: [Call]
      summary: Aggregate call stats (total, answered, missed, answer rate, ...)
      security:
        - bearerAuth: []
      parameters:
        - name: phoneNumberId
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Stats retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/CallStats" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /call/{id}:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Call]
      summary: Get a single call by id
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Call retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Call" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Call not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /call/{id}/recording:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Call]
      summary: Get a presigned download URL for this call's recording
      description: >-
        Only returns a URL once the recording has been downloaded from Meta and
        stored in MinIO (Fase 2). 404 if not requested/not ready yet, 410 if
        Meta's 7-day retention window passed before we could download it.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Presigned URL retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          url: { type: string }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: No recording, or not downloaded yet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "410":
          description: Expired — Meta deleted the media before we downloaded it
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /call/{id}/transcript:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Call]
      summary: Get this call's transcript content
      description: >-
        Same PENDING/EXPIRED lifecycle as /call/{id}/recording, but returns the
        parsed transcript JSON directly (Meta's documented shape: metadata,
        transcript.text/language/confidence, transcript.segments[] with
        speaker/channel/start/end/words) rather than a URL — the browser needs
        to actually read this, and a MinIO presigned URL may not have CORS
        enabled for direct fetch, so this goes through the authenticated API
        instead of the object store directly.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Transcript content retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        description: Meta's transcript document — see description above
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: No transcript, or not downloaded yet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "410":
          description: Expired — Meta deleted the media before we downloaded it
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ── Phone Number Admin ───────────────────────────────────────────────────
  /phone-number:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
    get:
      tags: [Phone Number]
      summary: Get all configured phone numbers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Phone numbers retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/PhoneNumber" }
                      meta: { $ref: "#/components/schemas/PaginationMeta" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /phone-number/{id}:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Phone Number]
      summary: Get a single phone number's configuration
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Phone number retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/PhoneNumber" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Phone number not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    put:
      tags: [Phone Number]
      summary: Update configuration and push the FULL config to Meta
      description: >-
        call_hours is a REPLACE on Meta's side, not a merge — this endpoint always sends
        the complete config. Changes can take up to 7 days to propagate to WhatsApp clients.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePhoneNumberRequest" }
      responses:
        "200":
          description: Updated and synced to Meta successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/PhoneNumber" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Phone number not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationError" }
        "502":
          description: Meta rejected the settings push (local save is kept regardless)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /phone-number/{id}/sync:
    parameters:
      - $ref: "#/components/parameters/AcceptLanguage"
      - name: id
        in: path
        required: true
        schema: { type: integer }
    post:
      tags: [Phone Number]
      summary: Re-push the current local config to Meta without changing it
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Synced successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/PhoneNumber" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "502":
          description: Meta rejected the settings push
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...